+++ lib/td_auth_web/views/group_view.ex
@@ -1,9 +1,14 @@
 defmodule TdAuthWeb.GroupView do
   use TdAuthWeb, :view
   alias TdAuthWeb.GroupView
+  alias TdAuthWeb.UserView
 
   def render("index.json", %{groups: groups}) do
     %{data: render_many(groups, GroupView, "group.json")}
+  end
+
+  def render("index.json", %{users: users}) do
+    %{data: render_many(users, UserView, "user_embedded.json")}
   end
 
   def render("show.json", %{group: group}) do
@@ -11,9 +16,7 @@ defmodule TdAuthWeb.GroupView do
   end
 
   def render("group.json", %{group: group}) do
-    %{id: group.id,
-      name: group.name,
-      description: group.description}
+    %{id: group.id, name: group.name, description: group.description}
   end
 
   def render("name.json", %{group: group}) do